# Convert portions of WordSeg to use scalars instead of Tensors