--- a/Examples/WordSeg/main.swift
+++ b/Examples/WordSeg/main.swift
@@ -78,7 +78,7 @@ for epoch in 1...maxEpochs {
   var trainingBatchCount = 0
   for record in dataset.training {
     let sentence = record.numericalizedText
-    let (loss, gradients) = valueWithGradient(at: model) { model -> Tensor<Float> in
+    let (loss, gradients) = valueWithGradient(at: model) { model -> Float in
       let lattice = model.buildLattice(sentence, maxLen: maxLength)
       let score = lattice[sentence.count].semiringScore
       let expectedLength = exp(score.logr - score.logp)
@@ -86,7 +86,7 @@ for epoch in 1...maxEpochs {
       return loss
     }
 
-    trainingLossSum += loss.scalarized()
+    trainingLossSum += loss
     trainingBatchCount += 1
     optimizer.update(&model, along: gradients)
 
@@ -132,7 +132,7 @@ for epoch in 1...maxEpochs {
     var lattice = model.buildLattice(sentence, maxLen: maxLength)
     let score = lattice[sentence.count].semiringScore
 
-    validationLossSum -= score.logp.scalarized()
+    validationLossSum -= score.logp
     validationBatchCount += 1
     validationCharacterCount += sentence.count