--- a/Models/Text/WordSeg/Model.swift
+++ b/Models/Text/WordSeg/Model.swift
@@ -334,3 +334,4 @@ extension Tensor {
     return (scalars, pullback)
   }
 }
+