--- a/Models/Text/WordSeg/SemiRing.swift
+++ b/Models/Text/WordSeg/SemiRing.swift
@@ -12,8 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import TensorFlow
-
 #if os(iOS) || os(macOS) || os(tvOS) || os(watchOS)
   import Darwin
 #elseif os(Windows)
@@ -26,17 +24,33 @@ import TensorFlow
 ///
 /// logSumExp (see https://en.wikipedia.org/wiki/LogSumExp)
 @differentiable
-public func logSumExp(_ x: [Tensor<Float>]) -> Tensor<Float> {
-  // Deal with an empty array first.
-  if x.count == 0 { return Tensor(-Float.infinity) }
-  return Tensor<Float>(stacking: x).logSumExp()
+public func logSumExp(_ x: [Float]) -> Float {
+  if x.count == 0 { return -Float.infinity}
+  let maxVal = x.max()!
+  let exps = x.map { exp($0 - maxVal) }
+  return maxVal + log(exps.reduce(into: 0, +=))
+}
+
+@derivative(of: logSumExp)
+public func vjpLogSumExp(_ x: [Float]) -> (
+  value: Float,
+  pullback: (Float) -> (Array<Float>.TangentVector)
+) {
+  func pb(v: Float) -> (Array<Float>.TangentVector) {
+    if x.count == 0 { return Array<Float>.TangentVector([]) }
+    let maxVal = x.max()!
+    let exps = x.map { exp($0 - maxVal) }
+    let sumExp = exps.reduce(into: 0, +=)
+    return Array<Float>.TangentVector(exps.map{ v * $0 / sumExp })
+  }
+  return (logSumExp(x), pb)
 }
 
 /// logSumExp(_:_:)
 ///
-/// Specialized logSumExp for 2 tensor of floats.
+/// Specialized logSumExp for 2 floats.
 @differentiable
-public func logSumExp(_ lhs: Tensor<Float>, _ rhs: Tensor<Float>) -> Tensor<Float> {
+public func logSumExp(_ lhs: Float, _ rhs: Float) -> Float {
   return logSumExp([lhs, rhs])
 }
 
@@ -44,19 +58,13 @@ public func logSumExp(_ lhs: Tensor<Float>, _ rhs: Tensor<Float>) -> Tensor<Floa
 ///
 /// Represents a SemiRing
 public struct SemiRing: Differentiable {
-  public var logp: Tensor<Float>
-  public var logr: Tensor<Float>
-
-  @differentiable
-  public init(logp: Tensor<Float>, logr: Tensor<Float>) {
-    self.logp = logp
-    self.logr = logr
-  }
+  public var logp: Float
+  public var logr: Float
 
   @differentiable
   public init(logp: Float, logr: Float) {
-    self.logp = Tensor(logp)
-    self.logr = Tensor(logr)
+    self.logp = logp
+    self.logr = logr
   }
 
   static var zero: SemiRing { SemiRing(logp: -Float.infinity, logr: -Float.infinity) }
@@ -97,10 +105,10 @@ extension SemiRing {
   // TODO(abdulras) see if we can use ulp as a default tolerance
   @inlinable
   public func isAlmostEqual(to other: Self, tolerance: Float) -> Bool {
-    return
-      (self.logp.isAlmostEqual(to: other.logp, tolerance: tolerance)
-      || (self.logp.scalarized().isInfinite && other.logp.scalarized().isInfinite))
-      && (self.logr.isAlmostEqual(to: other.logr, tolerance: tolerance)
-        || (self.logr.scalarized().isInfinite && other.logr.scalarized().isInfinite))
+    let diffP = abs(self.logp - other.logp)
+    let diffR = abs(self.logp - other.logp)
+
+    return (diffP <= tolerance || diffP.isNaN)
+      && (diffR <= tolerance || diffR.isNaN)
   }
 }