--- a/Tests/TextTests/WordSegmentationTests/ProbeLayers.swift
+++ b/Tests/TextTests/WordSegmentationTests/ProbeLayers.swift
@@ -200,7 +200,7 @@ class WordSegProbeLayerTests: XCTestCase {
     XCTAssert(lattice.isAlmostEqual(to: Example1.lattice, tolerance: 1e-5))
 
     print("Gradient")
-    func f(_ x: SNLM) -> Tensor<Float> {
+    func f(_ x: SNLM) -> Float {
       x.buildLattice(abab, maxLen: 5)[4].semiringScore.logr
     }
     let (_, grad) = valueWithGradient(at: model, in: f)