After profiling, it appears that we're creating a lot of small Tensors in the last steps of `buildLattice()`, where those calculations might better be performed using scalars on the host side. This PR therefore effectively rolls back portions of PR #535 to return to the use of scalars in many places.

Benchmarks before this change on a GTX 1080-equipped Ubuntu system:
WordSeg_score_4: 314 ms / step
WordSeg_score_and_gradient_4: 2801 ms / step

After:
WordSeg_score_4: 132 ms / step
WordSeg_score_and_gradient_4: 1246 ms / step